Move CubedPipeline to runtime package
