+++ cubed/runtime/executors/dask_distributed_async.py
@@ -18,10 +18,9 @@ from dask.distributed import Client
 from networkx import MultiDiGraph
 
 from cubed.core.array import Spec
-from cubed.primitive.types import CubedPipeline
 from cubed.runtime.executors.asyncio import async_map_unordered
 from cubed.runtime.pipeline import visit_node_generations, visit_nodes
-from cubed.runtime.types import Callback, DagExecutor
+from cubed.runtime.types import Callback, CubedPipeline, DagExecutor
 from cubed.runtime.utils import execution_stats, gensym, handle_callbacks
