+++ cubed/runtime/executors/python.py
@@ -3,9 +3,8 @@ from typing import Any, Callable, Optional, Sequence
 from networkx import MultiDiGraph
 
 from cubed.core.array import Spec
-from cubed.primitive.types import CubedPipeline
 from cubed.runtime.pipeline import visit_nodes
-from cubed.runtime.types import Callback, DagExecutor, TaskEndEvent
+from cubed.runtime.types import Callback, CubedPipeline, DagExecutor, TaskEndEvent
 
 
 def exec_stage_func(func: Callable[..., Any], *args, **kwargs):
