+++ cubed/runtime/executors/python_async.py
@@ -9,10 +9,9 @@ from networkx import MultiDiGraph
 from tenacity import Retrying, stop_after_attempt
 
 from cubed.core.array import Spec
-from cubed.primitive.types import CubedPipeline
 from cubed.runtime.executors.asyncio import async_map_unordered
 from cubed.runtime.pipeline import visit_node_generations, visit_nodes
-from cubed.runtime.types import Callback, DagExecutor
+from cubed.runtime.types import Callback, CubedPipeline, DagExecutor
 from cubed.runtime.utils import execution_stats, handle_callbacks
