+++ cubed/runtime/types.py
@@ -1,7 +1,10 @@
 from dataclasses import dataclass
-from typing import Optional
+from typing import Any, Iterable, Optional
 
 from networkx import MultiDiGraph
+
+from cubed.types import T_RegularChunks
+from cubed.vendor.rechunker.types import Config, StageFunction
 
 
 class DagExecutor:
@@ -10,6 +13,21 @@ class DagExecutor:
 
 
 Executor = DagExecutor
+
+
+@dataclass(frozen=True)
+class CubedPipeline:
+    """Generalisation of rechunker ``Pipeline`` with extra attributes."""
+
+    function: StageFunction
+    name: str
+    mappable: Iterable
+    config: Config
+    target_array: Any
+    projected_mem: int
+    reserved_mem: int
+    num_tasks: int
+    write_chunks: Optional[T_RegularChunks]
 
 
 class Callback:
